requirements.txt
Django==4.1.3
django-render-partial==0.4
djangorestframework==3.14.0
selenium==4.2.0
webdriver-manager==3.8.3
